==> notebooks/02-survey-expansion.Rmd <==
---
title: "Survey Expansion"
output: html_notebook
---

# Overhead
```{r overhead, include = FALSE}
packages_vector <- c("tidyverse", 
                     "yaml")
need_to_install <- packages_vector[!(packages_vector %in% installed.packages()[,"Package"])]

if (length(need_to_install)) install.packages(need_to_install)

for (package in packages_vector) {
  library(package, character.only = TRUE)
}

```

```{r wd}
print(getwd())
```

# Remote I/O
The user can change the expansion config in this block, see `config_file`. The other input files are specified in the `config_file`. 
```{r remote-io}
setwd(getwd())

processed_dir <- "../data/processed/"
interim_dir <- "../data/interim/"
output_dir <- "../data/interim/"

#config_file <- paste0(interim_dir, "expansion_config_departing_non_sas_only.yaml")
#config_file <- paste0(interim_dir, "expansion_config_departing_only.yaml")
#config_file <- paste0(interim_dir, "expansion_config_departing_and_arriving.yaml")
config_file <- paste0(interim_dir, "expansion_config_departing_only_with_time_of_day.yaml")


script_config <- yaml.load_file(config_file)

# Input I/O
survey_file <- paste0(processed_dir, script_config$input$survey_file)
controls_file <- paste0(interim_dir, script_config$input$controls_file)
config_file <- paste0(interim_dir, script_config$input$config_file)

# Output I/O
survey_expanded_file <- script_config$output$survey_expanded_file
convergence_stat_file <- script_config$output$stat_file
rfac_file <- script_config$output$rfac_file
weights_only_file <- script_config$output$weights_only_file

```

# Parameters
The parameters for the expansion:
1. `max_iteration` - The number of iterations to fit the data (50 should be more than enough)
2. `scaling_factor_low` - The minimum allowable weight is the mean expected weight (see `avgweight` in the `preparation` block below) divided by this factor
3. `scaling_factor_high` - The maximum allowable weight is the mean expected weight multiplied by this factor
4. `departing_only` - Set to `TRUE` for the departing only expansion and `FALSE` for the departing plus arriving expansion
```{r parameters}
max_iteration <- script_config$parameters$max_iteration
scaling_factor_low <- script_config$parameters$scaling_factor_low
scaling_factor_high <- script_config$parameters$scaling_factor_high
departing_only <- script_config$parameters$departing_only
non_sas_only <- script_config$parameters$non_sas_only

SURVEY_ID <- "unique_id"
```

# Reads
Read the data from disk, setting the data types for the key variables. 
```{r data-reads}
controls_df <- read_csv(controls_file, col_types = cols(.default = col_number(), airport_terminal = col_character()))
config_df <- read_csv(config_file, col_types = cols(.default = col_character(), priority = col_number(), use = col_integer()))
survey_df <- read_csv(survey_file, col_types = cols(.default = col_character(),
                                                    main_mode = col_number(),
                                                    marketsegment = col_number(),
                                                    inbound_or_outbound = col_number(),
                                                    party_size_flight = col_number(),
                                                    airport_terminal = col_number(),
                                                    flight_departure_time = col_number(),
                                                    weight = col_number()))
```

# Methods

## Create Incidence Table
For each row in the survey table, this method determines which of the controls are incident (i.e., relevant)
to the row. At least one control should be relevant to each control. 
```{r method-create-incidence-table}
create_incidence_table <- function(config_df, survey_df){

  incidence_df <- survey_df
  
  for (j in seq_len(nrow(config_df))) {
    j_config_df <- slice(config_df, j)
    var_name <- paste0(pull(j_config_df, controlid))
    expression <- pull(j_config_df, expression)
    
    incidence_df <- incidence_df %>%
      mutate(!!sym(var_name) := eval(parse(text = expression)))
  }

  return(incidence_df)
  
}
```

## Add Relaxation Factor to Controls
Initializes the relaxation controls used in the fitting algorithms to 1.0. 
```{r method-add-relaxation-factor}
add_rfac <- function(config_df, controls_df){

  controls_df <- controls_df %>%
    mutate(across(all_of(config_df$controlid), ~ 1, .names = "R_{.col}"))

  return(controls_df)
  
}
```

## Newton-Raphson Expansion
Uses the Netwon-Raphson method to fit the controls. For each control, it iteratively adjusts the weight to match the control.
The importance of each weight is use to dampen the correction. 
```{r method-expansion}
newton_raphson_expansion <- function(control_field, control_priority, s_df, c_df, index_key, control_key){
  
   hhagg_df <- s_df %>% 
      group_by(!!sym(index_key)) %>% 
      summarize(sum_value = eval(parse(text = gsub("x", control_field, "sum(x*weight, na.rm=TRUE)"))), 
                sum_value_sq = eval(parse(text = gsub("x", control_field, "sum(x*x*weight, na.rm=TRUE)")))) %>% 
      select(all_of(index_key), sum_value, sum_value_sq) %>% 
      ungroup()

   relax_factor <- paste0("R_", control_field)

   expression <- gsub("zz", control_priority, gsub("yy", relax_factor, gsub("xx", control_field, "ifelse(sum_value>0 & xx>0, 1- ((sum_value - xx*yy)/(sum_value_sq+ xx*yy/zz)), 1)")))
  
   update_controls_df <- c_df %>% 
     left_join(., hhagg_df, join_by(!!sym(control_key) == !!sym(index_key))) %>% 
     mutate(correction = eval(parse(text = expression))) %>% 
     select(-sum_value, -sum_value_sq)

   expression <- gsub("yy", control_field, gsub("zz", control_priority, gsub("xx", relax_factor, "ifelse( yy>0, xx*((1/correction)^(1/zz)), xx)")) ) 

   var_name <- paste0("R_", control_field)

   return_df <- update_controls_df %>% 
      mutate(!!sym(var_name) := eval(parse(text = expression))) 
  
   return(return_df)
  
}
```

## Expansion Statistics
Helper method to print out statistics for each round of expansion. 
```{r method-expansion-statistics}
run_stats <- function(s_df, c_df, control_field, iter) {
  
  max_weight <- max(s_df$weight, na.rm = TRUE)
  min_weight <- min(s_df$weight, na.rm = TRUE)
  mean_weight <- mean(s_df$weight, na.rm = TRUE)
  std_weight <- sd(s_df$weight, na.rm = TRUE)

  max_correction <- max(c_df$correction, na.rm = TRUE)
  min_correction <- min(c_df$correction, na.rm = TRUE)
  mean_correction <- mean(c_df$correction, na.rm = TRUE)
  std_correction <- sd(c_df$correction, na.rm = TRUE)

  r_df <- tibble(
    iter = iter, 
    control = control_field,
    mean_adj = mean_correction,
    min_adj = min_correction,
    max_adj = max_correction,
    std_adj = std_correction,
    mean_weight = mean_weight,
    min_weight = min_weight,
    max_weight = max_weight,
    sd_weight = std_weight
  )
  
  return(r_df)
}
```


# Preparation
Prepare the data files for the expansion, setting the incidence values for each row and initializing the relaxation factors.
```{r preparation}
working_controls_df <- controls_df %>%
  mutate(across(everything(), ~replace_na(., 0)))

working_config_df <- config_df %>%
  filter(use == 1L)

working_survey_df <- survey_df %>%
  filter(is_valid_record == "True") %>%
  mutate(count = 1) %>%
  mutate_all(~replace(., is.na(.), 0)) %>%
  mutate(party_size_flight = party_size_flight + 1) %>% 
  mutate(airport_terminal = paste0(as.integer(airport_terminal)))

if (departing_only) {
  working_survey_df <- working_survey_df %>%
    filter(inbound_or_outbound == 1L)
}


if (non_sas_only) {
  working_survey_df <- working_survey_df %>%
  filter(is_self_administered == "False")
}

survey_colnames = as.list(colnames(working_survey_df))

working_survey_df <- create_incidence_table(working_config_df, working_survey_df)

weights_df <- working_survey_df %>%
  group_by(marketsegment) %>%
  summarize(sum = sum(weight), .groups = "drop") %>%
  mutate(avgweight = ifelse(marketsegment == 2L, sum(working_controls_df$tot_emp)/sum, sum(working_controls_df$total_pax)/sum),
         minweight = avgweight/scaling_factor_low,
         maxweight = avgweight*scaling_factor_high)

working_survey_df <- left_join(working_survey_df, weights_df, by = c("marketsegment"))

working_controls_df <- add_rfac(working_config_df, working_controls_df)

statout_df <- tibble()
```

# Expansion
Run the expansion, which calls the fsitting algorithm `max_iteration` times. There is no assessment of fit, it just runs through `max_iteration`. Statistics for
each iteration are logged, so you can see how well the algorithm is doing. Runtime should be about 15 seconds on a laptop. 
```{r expansion}

running_controls_df <- working_controls_df
running_survey_df <- working_survey_df

for (iter in 1:max_iteration){
  print(iter)
  
  for (j in seq_len(nrow(working_config_df))){
    
    j_config_df <- slice(working_config_df, j) 

    control_field <- pull(j_config_df, "controlid")
    control_priority <- pull(j_config_df, "priority")
    index_key <- pull(j_config_df, "indexkey")
    control_key <- pull(j_config_df, "controlkey")

    if(!control_field %in% colnames(running_controls_df)) { print(paste("Control Not Found: ", control_field))}

    running_controls_df <- newton_raphson_expansion(control_field, control_priority, running_survey_df, running_controls_df, index_key, control_key)

    expression <- gsub("xx", control_field, "ifelse(xx>0, pmin(pmax(weight*(correction^xx),minweight),maxweight), weight)")

    running_survey_df <- running_survey_df %>%  
      left_join(., select(running_controls_df, !!sym(control_key), correction), join_by(!!sym(index_key) == !!sym(control_key))) %>% 
      mutate(weight = eval(parse(text = expression))) %>% 
      select(-correction)

    stats_df <- run_stats(running_survey_df, running_controls_df, control_field, iter)
 
    statout_df <- bind_rows(statout_df, stats_df)

  }

}
```


# Resolution
Prepare the results to be written to disk. 
```{r resolution}
survey_out_df <- running_survey_df %>% 
  select(match(survey_colnames, names(.))) %>% 
  select(-count)

relax_columns <- c(control_key, colnames(select(controls_df, starts_with("R_"))))

relax_out_df <- running_controls_df %>% 
  select(one_of(relax_columns)) 

weights_only_df <- survey_out_df %>% 
   select(!!sym(SURVEY_ID), weight)

```


# Writes
```{r writes}
write_csv(survey_out_df, file.path(interim_dir, survey_expanded_file))
write_csv(statout_df, file.path(interim_dir, convergence_stat_file)) 
write_csv(relax_out_df, file.path(interim_dir, rfac_file))
write_csv(weights_only_df, file.path(interim_dir, weights_only_file))
```
