==> notebooks/03-merge-weights-to-data.ipynb <==
import pandas as pd

data_model_output = pd.read_csv('../data/processed/data_model_output.csv')
weights_departing_and_arriving = pd.read_csv('../data/interim/survey_weights_only_departing_and_arriving.csv')
weights_departing_only = pd.read_csv('../data/interim/survey_weights_only_departing_only.csv')
weights_non_sas_departing_only = pd.read_csv('../data/interim/survey_weights_only_departing_non_sas_only.csv')
weights_departing_only_with_time_of_day = pd.read_csv('../data/interim/survey_weights_only_departing_only_with_time_of_day.csv')

# Drop the original weight column
data_model_output = data_model_output.drop(columns=['weight'], errors='ignore')

# Merge weights into the main dataset
data_model_output = data_model_output.merge(
    weights_departing_and_arriving[['unique_id', 'weight']], 
    on='unique_id', how='left'
).rename(columns={'weight': 'weight_departing_and_arriving'})

data_model_output = data_model_output.merge(
    weights_departing_only[['unique_id', 'weight']], 
    on='unique_id', how='left'
).rename(columns={'weight': 'weight_departing_only'})

data_model_output = data_model_output.merge(
    weights_non_sas_departing_only[['unique_id', 'weight']], 
    on='unique_id', how='left'
).rename(columns={'weight': 'weight_non_sas_departing_only'})

data_model_output = data_model_output.merge(
    weights_departing_only_with_time_of_day[['unique_id', 'weight']], 
    on='unique_id', how='left'
).rename(columns={'weight': 'weight_departing_only_with_time_of_day'})

data_model_output.head()

data_model_output.to_csv('../data/processed/data_model_output.csv', index=False)
