==> notebooks/06-mode-summaries-by-date.ipynb <==
import pandas as pd 
import numpy as np
import os

external_dir = "../data/external"
interim_dir = "../data/interim"
processed_dir = "../data/processed"


input_file_1 = os.path.join(external_dir, "etc/od_20241002_sandag_airport_pilot.xlsx")
input_file_2 = os.path.join(external_dir, "etc/od_20241004_sandag_airport_pilot_2.xlsx")
input_file_3 = os.path.join(external_dir, "etc/od_20241010_sandag_airport_pilot_3.xlsx") 
input_file = os.path.join(external_dir, "etc/od_20241015_sandag_airport_pilot_4.xlsx") #pilot survey 4, latest
variable_map_file = os.path.join(processed_dir, "revised_names.csv")

#in_df_1 = pd.read_excel(input_file_1)
#in_df_2 = pd.read_excel(input_file_2)
#in_df_3 = pd.read_excel(input_file_3)
#
#in_df = pd.concat([in_df_1, in_df_2, in_df_3], ignore_index = True)

in_df = pd.read_excel(input_file)

header_df = pd.read_csv(variable_map_file)[['ETC_name','WSP_name']]
header_dict = pd.Series(header_df.WSP_name.values,index=header_df.ETC_name).to_dict()
clean_df = in_df.rename(columns=header_dict).copy().drop(columns=["delete"])

clean_df.shape

clean_df.head()

#Remove duplicate respondentid
clean_df = clean_df.drop_duplicates(subset='respondentid', keep='first')
clean_df.reset_index(drop=True, inplace=True)
clean_df.shape

cols_with_oth = [col for col in clean_df.columns if '-oth-' in clean_df[col].values]
cols_with_oth

cols_with_oth_and_98 = [col for col in clean_df[cols_with_oth] if 98 in clean_df[col].values]
cols_with_oth_and_98

def replace_oth_with_98(df):
    # Iterate over each mode column
    for col in df.columns:
        # Replace '-oth-' with 98
        df[col] = df[col].replace('-oth-', 98)
    
    return df

clean_df = replace_oth_with_98(clean_df)

### Mode summarization starts here

#Select columns relevant to mode
exclude_substrings = ('general', 'alt', 'split', 'same', 'sdia', '_other')
mode_columns = [col for col in clean_df.columns if 'mode' in col and all(sub not in col for sub in exclude_substrings)]
print(mode_columns)

mode_code_columns = [col for col in mode_columns if 'label' not in col]
mode_code_columns

mode_label_columns = [col for col in mode_columns if 'label' in col]
mode_label_columns

value_counts_df = pd.concat([clean_df[col].value_counts().rename(col) for col in mode_columns], axis=1)

print(value_counts_df)

clean_df['main_mode'].value_counts()

clean_df['main_mode_label'].value_counts()

clean_df['main_transit_mode'].value_counts()

clean_df['main_transit_mode_label'].value_counts()

clean_df['date_completed'] = pd.to_datetime(clean_df['date_completed'])

unique_dates = clean_df['date_completed'].unique()
unique_dates

import pandas as pd

def get_value_counts_with_labels(df, mode_columns, label_columns):
    # Create an empty list to store the result
    result = []

    # Iterate over each pair of code and label columns
    for col, label_col in zip(mode_columns, label_columns):
        # Get value counts for the code column
        counts = df[col].value_counts().reset_index()
        counts['share'] = counts['count']/np.sum(counts['count'])

        # Get value counts for the label column
        label_counts = df[label_col].value_counts().reset_index()
        label_counts['share'] = label_counts['count']/np.sum(label_counts['count'])

        # Rename columns to 'value' and 'count'
        counts.columns = ['value', 'count', 'count_share']
        label_counts.columns = ['label_value', 'label_count', 'label_share']

        # Merge both DataFrames side by side (assuming they match by index)
        merged = pd.concat([counts, label_counts], axis=1)

        # Add a column for the column name (code column)
        merged['column'] = col

        # Append the result to the list
        result.append(merged)

    # Concatenate all the DataFrames in the list into a single DataFrame
    final_df = pd.concat(result, ignore_index=True)

    # Reorder columns to have 'column', 'value', 'count', 'label_value', 'label_count'
    final_df = final_df[['column', 'value', 'label_value', 'count', 'count_share', 'label_count', 'label_share']]

    return final_df


value_counts_df = get_value_counts_with_labels(clean_df, mode_code_columns, mode_label_columns)
print(value_counts_df)

clean_df['date_completed']

value_counts_df.to_csv("../data/processed/Mode_summaries.csv", index = False)

clean_df['access_mode'].value_counts()

clean_df['access_mode_label'].value_counts()

clean_df['reverse_mode_predicted'].value_counts()

clean_df['reverse_mode_predicted_label'].value_counts()

#main_mode code = 16 label and code don't exist
#access_mode, reverse_mode_predicted 

unique_dates = clean_df['date_completed'].unique()
unique_dates

total_df = pd.DataFrame()
for i in unique_dates:
    df_by_date = clean_df[clean_df['date_completed'] == i]
    value_counts_df = get_value_counts_with_labels(df_by_date, mode_code_columns, mode_label_columns)
    value_counts_df['date'] = i
    total_df = pd.concat([total_df,value_counts_df], ignore_index = True)


total_df.head()

total_df.to_csv("../reports/Mode_summaries_by_date.csv", index = False)
