==> notebooks/14-filter-data-model-complete-records.ipynb <==
import pandas as pd  
import numpy as np  

data_model_file_name = '../data/processed/data_model_output.csv'
clean_output_file_name = '../data/processed/atc_travel_survey_final_data.csv'
column_filter_file_name = '../data/processed/columns_to_keep.csv'

data_model_output = pd.read_csv(data_model_file_name)
column_filter = pd.read_csv(column_filter_file_name)

# Step 1: Filter rows where validation_severity_person and validation_severity_trip are NOT 'Critical'
# and is_completed is True
filtered_df = data_model_output[data_model_output['is_valid_record'] == True]

# Step 2: Apply column filter logic
# Create a mapping from column_filter where 'keep' means keep, 'delete' means remove, and others mean rename
column_mapping = dict(zip(column_filter["column"], column_filter["status"]))

# Step 3: Drop columns marked as 'delete'
columns_to_drop = [col for col, status in column_mapping.items() if status == "delete"]
filtered_df = filtered_df.drop(columns=columns_to_drop, errors="ignore")

# Step 4: Rename columns that have a different mapping
columns_to_rename = {col: new_name for col, new_name in column_mapping.items() if new_name not in ["keep", "delete"]}
filtered_df = filtered_df.rename(columns=columns_to_rename)



filtered_df.shape, data_model_output.shape

# Step 5: Save the cleaned data
filtered_df.to_csv(clean_output_file_name, index=False)
