==> notebooks/17-business-vs-non-business-logic.ipynb <==
import pandas as pd
import numpy as np
import os
import matplotlib.pyplot as plt

interim_dir = "../data/interim"
processed_dir = "../data/processed"

data_model_file = os.path.join(processed_dir, "data_model_output.csv")

data_model_df = pd.read_csv(data_model_file)
data_model_df.head()
data_model_df = data_model_df[(data_model_df['is_completed'] == 1) & (data_model_df['record_type_synthetic'] == 0) & (data_model_df['submit']==True)]

working_df = data_model_df[['marketsegment_label', 'passenger_segment_label', 'flight_purpose_label', 'is_sdia_home_airport','reimbursement_label']]

working_df.head()

# Define the classification logic using np.where
working_df['passenger_flight_purpose_segment'] = np.where(
    working_df['is_sdia_home_airport'] & working_df['flight_purpose_label'].isin(['BUSINESS_WORK', 'COMBINATION_BUSINESS_LEISURE']),
    'RESIDENT_BUSINESS',
    np.where(
        working_df['is_sdia_home_airport'], 'RESIDENT_NON_BUSINESS',
        np.where(
            working_df['flight_purpose_label'].isin(['BUSINESS_WORK', 'COMBINATION_BUSINESS_LEISURE']),
            'VISITOR_BUSINESS',
            'VISITOR_NON_BUSINESS'
        )
    )
)

passengers_df = working_df[working_df['marketsegment_label']=='PASSENGER']

passengers_df.head()

passengers_df['passenger_flight_purpose_segment'].value_counts()

passengers_df['reimbursement_label'].value_counts()

passengers_df['flight_purpose_label'].value_counts()

passengers_df[passengers_df['flight_purpose_label']=='COMMUTE']

passengers_df[['passenger_flight_purpose_segment', 'reimbursement_label']].groupby(['passenger_flight_purpose_segment', 'reimbursement_label']).size()

crosstab_df = pd.crosstab(passengers_df['reimbursement_label'], passengers_df['passenger_flight_purpose_segment'], margins=True, margins_name='Total')
crosstab_percentage = crosstab_df.div(crosstab_df.loc['Total'], axis=1) * 100  # Column-wise percentage

# Concatenate count and percentage crosstab
crosstab_combined = crosstab_df.astype(str) + ' (' + crosstab_percentage.round(2).astype(str) + '%)'

# Display the final combined crosstab
crosstab_combined

crosstab_df = pd.crosstab(passengers_df['reimbursement_label'], passengers_df['flight_purpose_label'], margins=True, margins_name='Total')
crosstab_percentage = crosstab_df.div(crosstab_df.loc['Total'], axis=1) * 100  # Column-wise percentage

# Concatenate count and percentage crosstab
crosstab_combined = crosstab_df.astype(str) + ' (' + crosstab_percentage.round(2).astype(str) + '%)'

# Display the final combined crosstab
crosstab_combined
