==> notebooks/19-make-tall-data-for-select-tableau-summaries.ipynb <==
import pandas as pd
import numpy as np
import os

interim_dir = "../data/interim"
processed_dir = "../data/processed"
data_model_file = os.path.join(processed_dir, "data_model_output.csv")

output_filename = os.path.join(processed_dir, "data_model_output_select_tall_summaries.csv")

data_model_df = pd.read_csv(data_model_file)
df = data_model_df[(data_model_df['is_completed'] == 1) & (data_model_df['record_type_synthetic'] == 0) & (data_model_df['initial_etc_check']==True) & (data_model_df['marketsegment_label']=='PASSENGER')]

sp_feature_df = df.filter(like="sp_feature", axis=1)
sp_feature_df.columns

reason_df = df.filter(like="reason", axis=1) 
reason_df.columns

weight_df = df.filter(like="weight", axis=1)
weight_df = pd.concat([df["unique_id"], weight_df], axis=1)
weight_df.columns

working_df = pd.concat([df[["unique_id"]], reason_df, sp_feature_df], axis=1)
working_df.head()

tall_df = pd.melt(working_df, id_vars=["unique_id"], var_name="question", value_name="answer")
tall_df["question_type"] = "Missing"
tall_df["question_type"] = np.where(tall_df["question"].str.contains("reasons_no_transit"), "Reasons No Transit", tall_df["question_type"])
tall_df["question_type"] = np.where(tall_df["question"].str.contains("sp_feature"), "SP Feature", tall_df["question_type"])
tall_df["question"] = tall_df["question"].str.replace("reasons_no_transit_", "").str.replace("sp_feature_", "").str.replace("_"," ").str.title()
tall_df = tall_df.dropna(subset=["answer"]).reset_index(drop=True)
tall_df.head()

output_df = pd.merge(tall_df, weight_df, on="unique_id", how="left")
output_df.head()

output_df.to_csv(output_filename, index=False)
